+++ websocket/eslint.config.js
@@ -1,0 +1,21 @@
+const typescript = require("@typescript-eslint/eslint-plugin");
+const tsParser = require("@typescript-eslint/parser");
+
+module.exports = [
+  {
+    // test.js temporarily, unless it's adjusted and ported to TypeScript:
+    ignores: ["node_modules", "dist", "test.js"]
+  },
+  {
+    files: ["**/*.ts"],
+    languageOptions: {
+      parser: tsParser,
+      parserOptions: {
+        project: "./tsconfig.json"
+      }
+    },
+    plugins: {
+      "@typescript-eslint": typescript
+    },
+  },
+];